+++ src/main/java/ru/gadjini/telegram/smart/bot/commons/command/impl/LanguageCommand.java
@@ -79,7 +79,7 @@ public class LanguageCommand implements KeyboardBotCommand, BotCommand, Navigabl
         return true;
     }
 
-    private void processMessage0(long chatId, int userId) {
+    private void processMessage0(long chatId, long userId) {
         Locale locale = userService.getLocaleOrDefault(userId);
         messageService.sendMessage(SendMessage.builder().chatId(String.valueOf(chatId))
                 .text(localisationService.getMessage(MessagesProperties.MESSAGE_CHOOSE_LANGUAGE, locale))
