+++ src/main/java/ru/gadjini/telegram/smart/bot/commons/service/message/MessageService.java
@@ -18,7 +18,7 @@ public interface MessageService {
 
     void sendAnswerCallbackQuery(AnswerCallbackQuery answerCallbackQuery);
 
-    boolean isChatMember(String chatId, int userId);
+    boolean isChatMember(String chatId, long userId);
 
     void sendMessage(SendMessage sendMessage);
