+++ src/main/java/ru/gadjini/telegram/smart/bot/commons/service/message/TgLimitsMessageService.java
@@ -43,7 +43,7 @@ public class TgLimitsMessageService implements MessageService {
     }
 
     @Override
-    public boolean isChatMember(String chatId, int userId) {
+    public boolean isChatMember(String chatId, long userId) {
         return messageService.isChatMember(chatId, userId);
     }
